Create callback to notify when user session has changed
- SessionListener is the interface to implement in order to receive
callbacks from Digits when the session changes.

- New methods Digits#addSessionListener &
Digits#removeSessionListener. Those methods allow the developer to add
and remove more than one SessionListener.
This Allow developers to add and remove listeners during Activity lifecycle.

The recommend way to use this new feature is to add a listener on
Activity#onCreate and remove that listener on Activity#onStop

Change-Id: Id8ae1873d6b6c95a840cec4f27b7380431579829